+++ src/qutip_qoc/__init__.py
@@ -1,8 +1,8 @@
 from .version import version as __version__  # noqa
 
-from qutip_qoc.time import TimeInterval
+from qutip_qoc._time import _TimeInterval
 from qutip_qoc.result import Result
 from qutip_qoc.objective import Objective
 from qutip_qoc.pulse_optim import optimize_pulses
 
-__all__ = ["TimeInterval", "Result", "Objective", "optimize_pulses"]
+__all__ = ["_TimeInterval", "Result", "Objective", "optimize_pulses"]